+++ src/components/common/CollapseCard.vue
@@ -12,12 +12,12 @@
       />
     </div>
     <div
-      class="collapse-content p-0"
+      class="collapse-content !p-0"
       @transitionend="handlerTransitionEnd"
     >
       <div
         v-if="shouldRenderContent"
-        class="p-4 pt-0 max-md:px-3 max-md:py-2.5"
+        class="px-4 pt-0 pb-2.5 max-md:px-3 max-md:pt-0 max-md:pb-2.5"
         :class="[
           contentScrollable && 'max-h-108 overflow-y-auto',
           contentScrollable && SCROLLABLE_PARENT_CLASS,